+++ src/popups/modals/replace-translation.tsx
@@ -10,8 +10,13 @@ import { Switch } from "@/components/ui/switch";
 import { Input } from "@/components/ui/input";
 import { PopupFormProps } from "@/lib/types";
 import TranslatorActions from "@/actions/translator";
+import { useAppTranslation } from "@/context/translation";
+import { useState } from "react";
+import { toast } from "sonner";
 
 export default function ReplaceTranslationPopup({triggerButton}: PopupFormProps){
+     const {table, setTable} = useAppTranslation()
+     const [open, setOpen] = useState(false)
      const form = useForm<ReplaceTranslationType>({
           resolver: zodResolver(ReplaceTranslationSchema),
           defaultValues: {
@@ -21,7 +26,16 @@ export default function ReplaceTranslationPopup({triggerButton}: PopupFormProps)
           }
      })
      const onSubmit = (values: ReplaceTranslationType) => {
-          TranslatorActions.replaceTranslation(values)
+          const res = TranslatorActions.replaceTranslation(values,table)
+          if(res.error) toast.error("Failed to replace translations",{
+               description: res.error
+          })
+          if(res.success) {
+               toast.success(res.success)
+               setOpen(false)
+               setTable(res.data)
+               form.reset()
+          }
      }
      return (
           <AppModal
@@ -29,6 +43,7 @@ export default function ReplaceTranslationPopup({triggerButton}: PopupFormProps)
                title="Replace the translation"
                description="Replace existing translations"
                triggerButton={triggerButton}
+               open={open} onOpenChange={setOpen}
           >
                <form id="replace" onSubmit={form.handleSubmit(onSubmit)}>
                     <FieldGroup>
@@ -43,24 +58,6 @@ export default function ReplaceTranslationPopup({triggerButton}: PopupFormProps)
                                              id={field.name}
                                              aria-invalid={fieldState.invalid}
                                              placeholder="Cancel"
-                                        />
-                                        {fieldState.invalid && (
-                                             <FieldError errors={[fieldState.error]} />
-                                        )}
-                                   </Field>
-                              )}
-                         />
-                         <Controller
-                              control={form.control}
-                              name="to"
-                              render={({field, fieldState})=>(
-                                   <Field data-invalid={fieldState.invalid}>
-                                        <FieldLabel htmlFor={field.name}>To</FieldLabel>
-                                        <Input
-                                             {...field}
-                                             id={field.name}
-                                             aria-invalid={fieldState.invalid}
-                                             placeholder="Չեղարկել"
                                         />
                                         {fieldState.invalid && (
                                              <FieldError errors={[fieldState.error]} />
@@ -91,6 +88,24 @@ export default function ReplaceTranslationPopup({triggerButton}: PopupFormProps)
                                              onCheckedChange={field.onChange}
                                              aria-invalid={fieldState.invalid}
                                         />
+                                   </Field>
+                              )}
+                         />
+                         <Controller
+                              control={form.control}
+                              name="to"
+                              render={({field, fieldState})=>(
+                                   <Field data-invalid={fieldState.invalid}>
+                                        <FieldLabel htmlFor={field.name}>To</FieldLabel>
+                                        <Input
+                                             {...field}
+                                             id={field.name}
+                                             aria-invalid={fieldState.invalid}
+                                             placeholder="Չեղարկել"
+                                        />
+                                        {fieldState.invalid && (
+                                             <FieldError errors={[fieldState.error]} />
+                                        )}
                                    </Field>
                               )}
                          />